+++ app/src/main/java/elr/calculatortest1/io/SevenInputHandlerImpl.java
@@ -9,7 +9,7 @@ public class SevenInputHandlerImpl extends NumberInputHandler {
     }
 
     @Override
-    protected double getInputValue() {
-        return 7.0;
+    protected int getInputValue() {
+        return 7;
     }
 }
